maj is granted

+++ src/Controller/ArticleController.php
@@ -25,8 +25,10 @@ class ArticleController extends AbstractController
     }
 
     #[Route('/new', name: 'app_article_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, ArticleRepository $articleRepository, SluggerInterface $slugger): Response
+   
+    public function  new(Request $request, ArticleRepository $articleRepository, SluggerInterface $slugger): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_EDITOR", null, "Vous n'avez pas le droit d'accéder a cette page");
         $article = new Article();
         $form = $this->createForm(ArticleType::class, $article);
         $form->handleRequest($request);
@@ -78,6 +80,7 @@ class ArticleController extends AbstractController
     #[Route('/{id}/edit', name: 'app_article_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Article $article, ArticleRepository $articleRepository): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_EDITOR", null, "Vous n'avez pas le droit d'accéder a cette page");
         $form = $this->createForm(ArticleType::class, $article);
         $form->handleRequest($request);
 
@@ -96,6 +99,7 @@ class ArticleController extends AbstractController
     #[Route('/{id}', name: 'app_article_delete', methods: ['POST'])]
     public function delete(Request $request, Article $article, ArticleRepository $articleRepository): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_EDITOR", null, "Vous n'avez pas le droit d'accéder a cette page");
         if ($this->isCsrfTokenValid('delete' . $article->getId(), $request->request->get('_token'))) {
             $articleRepository->remove($article, true);
         }
