+++ src/Controller/ContactController.php
@@ -16,6 +16,7 @@ class ContactController extends AbstractController
     #[Route('/', name: 'app_contact_index', methods: ['GET'])]
     public function index(ContactRepository $contactRepository): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_EDITOR", null, "Vous n'avez pas le droit d'accéder a cette page");
         return $this->render('contact/index.html.twig', [
             'contacts' => $contactRepository->findAll(),
         ]);
@@ -43,6 +44,7 @@ class ContactController extends AbstractController
     #[Route('/{id}', name: 'app_contact_show', methods: ['GET'])]
     public function show(Contact $contact): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_EDITOR", null, "Vous n'avez pas le droit d'accéder a cette page");
         return $this->render('contact/show.html.twig', [
             'contact' => $contact,
         ]);
@@ -51,6 +53,7 @@ class ContactController extends AbstractController
     #[Route('/{id}/edit', name: 'app_contact_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Contact $contact, ContactRepository $contactRepository): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_EDITOR", null, "Vous n'avez pas le droit d'accéder a cette page");
         $form = $this->createForm(ContactType::class, $contact);
         $form->handleRequest($request);
 
@@ -69,6 +72,7 @@ class ContactController extends AbstractController
     #[Route('/{id}', name: 'app_contact_delete', methods: ['POST'])]
     public function delete(Request $request, Contact $contact, ContactRepository $contactRepository): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_EDITOR", null, "Vous n'avez pas le droit d'accéder a cette page");
         if ($this->isCsrfTokenValid('delete'.$contact->getId(), $request->request->get('_token'))) {
             $contactRepository->remove($contact, true);
         }
