+++ src/Controller/SocialNetworkController.php
@@ -16,6 +16,7 @@ class SocialNetworkController extends AbstractController
     #[Route('/', name: 'app_social_network_index', methods: ['GET'])]
     public function index(SocialNetworkRepository $socialNetworkRepository): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_EDITOR", null, "Vous n'avez pas le droit d'accéder a cette page");
         return $this->render('social_network/index.html.twig', [
             'social_networks' => $socialNetworkRepository->findAll(),
         ]);
@@ -24,6 +25,7 @@ class SocialNetworkController extends AbstractController
     #[Route('/new', name: 'app_social_network_new', methods: ['GET', 'POST'])]
     public function new(Request $request, SocialNetworkRepository $socialNetworkRepository): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_EDITOR", null, "Vous n'avez pas le droit d'accéder a cette page");
         $socialNetwork = new SocialNetwork();
         $form = $this->createForm(SocialNetworkType::class, $socialNetwork);
         $form->handleRequest($request);
@@ -43,6 +45,7 @@ class SocialNetworkController extends AbstractController
     #[Route('/{id}', name: 'app_social_network_show', methods: ['GET'])]
     public function show(SocialNetwork $socialNetwork): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_EDITOR", null, "Vous n'avez pas le droit d'accéder a cette page");
         return $this->render('social_network/show.html.twig', [
             'social_network' => $socialNetwork,
         ]);
@@ -51,6 +54,7 @@ class SocialNetworkController extends AbstractController
     #[Route('/{id}/edit', name: 'app_social_network_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, SocialNetwork $socialNetwork, SocialNetworkRepository $socialNetworkRepository): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_EDITOR", null, "Vous n'avez pas le droit d'accéder a cette page");
         $form = $this->createForm(SocialNetworkType::class, $socialNetwork);
         $form->handleRequest($request);
 
@@ -69,6 +73,7 @@ class SocialNetworkController extends AbstractController
     #[Route('/{id}', name: 'app_social_network_delete', methods: ['POST'])]
     public function delete(Request $request, SocialNetwork $socialNetwork, SocialNetworkRepository $socialNetworkRepository): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_EDITOR", null, "Vous n'avez pas le droit d'accéder a cette page");
         if ($this->isCsrfTokenValid('delete'.$socialNetwork->getId(), $request->request->get('_token'))) {
             $socialNetworkRepository->remove($socialNetwork, true);
         }
